polymarket_smart_copy_bot/requirements.txt
aiohttp==3.12.15
aiosqlite==0.20.0
alembic==1.14.1
apscheduler==3.10.4
asyncpg==0.30.0
certifi==2026.2.25
fastapi==0.115.8
greenlet==3.1.1
httpx==0.28.1
loguru==0.7.3
orjson==3.10.15
pydantic==2.10.6
pydantic-settings==2.8.1
psycopg2-binary==2.9.10
py-clob-client==0.34.6
PyYAML==6.0.2
python-dotenv==1.0.1
python-telegram-bot==21.10
rich==13.9.4
SQLAlchemy==2.0.38
tenacity==9.0.0
typer==0.15.1
uvicorn[standard]==0.34.0

pytest==8.3.4
pytest-asyncio==0.25.3
